Reblocks API change requires now 4-arguments for initiateFormAction
The Reblocks API change that required the 4-argument
initiateFormAction call is was introduced and is documented in
Reblocks PR #57: https://github.com/40ants/reblocks/pull/57

This PR introduced changes to the initiateFormAction JavaScript
function that required updating the function signature from 3
arguments to 4 arguments, where the fourth argument is an options
object.

The current fix in reblocks-auth was specifically implemented to
accommodate this breaking change in Reblocks core framework.